% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cols_omv.R
\name{combine_cols_omv}
\alias{combine_cols_omv}
\title{Combines pairs of columns from a raw data matrix in .omv-files for the statistical spreadsheet
'jamovi' (\url{https://www.jamovi.org})}
\usage{
combine_cols_omv(
  dtaInp = NULL,
  fleOut = "",
  varPrs = list(),
  usePkg = c("foreign", "haven"),
  selSet = "",
  ...
)
}
\arguments{
\item{dtaInp}{Either a data frame or the name of a data file to be read (including the path,
if required; "FILENAME.ext"; default: NULL); files can be of any supported file
type, see Details below.}

\item{fleOut}{Name of the data file to be written (including the path, if required;
"FILE_OUT.omv"; default: ""); if empty, the resulting data frame is returned
instead.}

\item{varPrs}{Variable (default: list()) containing a list of lists where each nested list
contains the names of pairs of variables to be combined.}

\item{usePkg}{Name of the package: "foreign" or "haven" that shall be used to read SPSS, Stata,
and SAS files; "foreign" is the default (it comes with base R), but "haven" is
newer and more comprehensive.}

\item{selSet}{Name of the data set that is to be selected from the workspace (only applies when
reading .RData-files)}

\item{...}{Additional arguments passed on to methods; see Details below.}
}
\value{
a data frame containing the column pairs given in \code{varPrs} combined and the original
columns removed
}
\description{
Combines pairs of columns from a raw data matrix in .omv-files for the statistical spreadsheet
'jamovi' (\url{https://www.jamovi.org})
}
\details{
\itemize{
\item 
\item The ellipsis-parameter (\code{...}) can be used to submit arguments / parameters to the functions
that are used for reading and writing the data. By clicking on the respective function under
“See also”, you can get a more detailed overview over which parameters each of those functions
take. The functions are: \code{read_omv} and \code{write_omv} (for jamovi-files), \code{read.table} (for CSV
/ TSV files; using similar defaults as \code{read.csv} for CSV and \code{read.delim} for TSV which both
are based upon \code{read.table}), \code{load} (for .RData-files), \code{readRDS} (for .rds-files),
\code{read_sav} (needs the R-package \code{haven}) or \code{read.spss} (needs the R-package \code{foreign}) for
SPSS-files, \code{read_dta} (\code{haven}) / \code{read.dta} (\code{foreign}) for Stata-files, \code{read_sas}
(\code{haven}) for SAS-data-files, and \code{read_xpt} (\code{haven}) / \code{read.xport} (\code{foreign}) for
SAS-transport-files. If you would like to use \code{haven}, you may need to install it using
\code{install.packages("haven", dep = TRUE)}.
}
}
\examples{
\dontrun{

}

}
\seealso{
\code{combine_cols_omv} uses the following functions for reading and writing data files in
different formats: \code{\link[=read_omv]{read_omv()}} and \code{\link[=write_omv]{write_omv()}} for
jamovi-files, \code{\link[utils:read.table]{utils::read.table()}} for CSV / TSV files, \code{\link[=load]{load()}} for reading .RData-files,
\code{\link[=readRDS]{readRDS()}} for .rds-files, \code{\link[haven:read_spss]{haven::read_sav()}} or \code{\link[foreign:read.spss]{foreign::read.spss()}} for SPSS-files,
\code{\link[haven:read_dta]{haven::read_dta()}} or \code{\link[foreign:read.dta]{foreign::read.dta()}} for Stata-files, \code{\link[haven:read_sas]{haven::read_sas()}} for
SAS-data-files, and \code{\link[haven:read_xpt]{haven::read_xpt()}} or \code{\link[foreign:read.xport]{foreign::read.xport()}} for SAS-transport-files.
}
