% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmvReadWrite.R
\docType{package}
\name{jmvReadWrite}
\alias{jmvReadWrite}
\title{jmvReadWrite: A package to read and write '.omv'-files (produced by the
statistical spreadsheet 'jamovi', www.jamovi.org).}
\description{
The jmvReadWrite package provides two functions:
jmvRead and jmvWrite.
}
\section{jmvReadWrite functions}{

The free and open a statistical spreadsheet 'jamovi' (www.jamovi.org) aims
<<<<<<< HEAD
to make statistical analyses easy and intuitive. 'jamovi' produces syntax
=======
to makes statistical analyses easy and intuitive. 'jamovi' produces syntax
>>>>>>> 92f8ecbd43221756663a1963418fae033e9cd09b
that can directly be used in R (in connection with the R-package 'jmv').
Having import / export routines for the data files generated by 'jamovi'
('.omv') permits an easy transfer of analyses between 'jamovi' and R.
}

