% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmvWrite.R
\name{jmvWrite}
\alias{jmvWrite}
\title{Write files to be used in jamovi (statistical spreadsheet, www.jamovi.org)}
\usage{
jmvWrite(dtaFrm = NULL, fleNme = "")
}
\arguments{
\item{dtaFrm}{Data frame to be exported (default = NULL)}

\item{fleNme}{Name / position of the output file to be generated (.omv; default = "")}
}
\description{
Write files to be used in jamovi (statistical spreadsheet, www.jamovi.org)
}
