% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cols_omv.R
\name{merge_cols_omv}
\alias{merge_cols_omv}
\title{Merges two or more data files by adding the content of other input files as columns to the first input file and outputs them as files for the statistical spreadsheet 'jamovi' (www.jamovi.org)}
\usage{
merge_cols_omv(
  fleInp = c(),
  fleOut = "",
  typMrg = c("outer", "inner", "left", "right"),
  varBy = list(),
  varSrt = c(),
  usePkg = c("haven", "foreign"),
  selSet = "",
  ...
)
}
\arguments{
\item{fleInp}{vector with the file names of the input files (including the path, if required; c('FILE_IN1.omv', 'FILE_IN2.omv'); default: c())}

\item{fleOut}{Name of the data file to be written (including the path, if required; 'FILE_OUT.omv'; default: ''); if empty, the data frame with the added columns is returned as variable (but not written)}

\item{typMrg}{Type of merging operation: 'inner' (default), 'outer', 'left' or 'right'; see Details below}

\item{varBy}{Name of the variable by which the data sets are matched, can either be a string, a character or a list (see Details below; default: list())}

\item{varSrt}{Variable(s) that are used to sort the data frame (see Details; if empty, the order after merging is kept; default: c())}

\item{usePkg}{Name of the package: 'haven' or 'foreign' that shall be used to read SPSS, Stata and SAS files; 'haven' is the default (it is more comprehensive), but with problems you may try 'foreign'}

\item{selSet}{Name of the data set that is to be selected from the workspace (only applies when reading .RData-files)}

\item{...}{Additional arguments passed on to methods; see Details below}
}
\value{
a data frame (if fleOut is empty) with where the columns of all input data sets (in the files given to fleInp) are concatenated
}
\description{
Merges two or more data files by adding the content of other input files as columns to the first input file and outputs them as files for the statistical spreadsheet 'jamovi' (www.jamovi.org)
}
\details{
There are four different types of merging operations: 'outer' keeps all cases (but columns in the resulting data set may be empty if they did not contain values in same input data sets), 'inner' keeps
only those cases where all datasets contain the same value in the matching variable, for 'left' all cases from the first data set in fleInp are kept (whereas cases that are only contained in input data
set two or higher are dropped), for 'right' all cases from the second (or any higher) data set in fleInp are kept. The behaviour of 'left' and 'right' may be somewhat difficult to predict in case of
merging several data sets, therefore 'outer' might be a safer choice if several data sets are merged.
The variable that is used for matching (varBy) can either be a string (if all datasets contain a matching variable with the same name), a character vector (containing several matching variables that
are the same for all data sets) or a list with the same length as fleInp. In the latter case, each cell of that list can again contain either a string (one matching variable for each data set in fleInp)
or a character vector (several matching variables for each data set in fleInp; NB: all character vectors in the cells of the list must have the same length as it is necessary to always use the same
number of matching variables when merging).
The ellipsis-parameter (...) can be used to submit arguments / parameters to the functions that are used for merging or reading the data. Adding columns uses 'merge'. When reading the data, the
functions are: 'read_omv' (for jamovi-files), 'read.table' (for CSV / TSV files; using similar defaults as 'read.csv' for CSV and 'read.delim' for TSV which both are based upon 'read.table' but
with adjusted defaults for the respective file types), 'readRDS' (for rds-files), 'read_sav' (needs R-package 'haven') or 'read.spss' (needs R-package 'foreign') for SPSS-files, 'read_dta'
('haven') / 'read.dta' ('foreign') for Stata-files, 'read_sas' ('haven') for SAS-data-files, and 'read_xpt' ('haven') / 'read.xport' ('foreign') for SAS-transport-files.
Please note that the R-packages 'haven' and 'foreign' are not marked as 'Imports' (i.e., they are not installed by default). If you wish to convert files from SPSS, SAS or Stata and haven't installed
them yet, please install them manually (e.g., \code{install.packages('haven', dep = TRUE)}).
}
\examples{
\dontrun{
library(jmvReadWrite);

}

}
