% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows_omv.R
\name{merge_rows_omv}
\alias{merge_rows_omv}
\title{Merges two .omv-files for the statistical spreadsheet 'jamovi' (www.jamovi.org) by adding the content of the second file (fleIn2) as rows to the first file (fleIn1)}
\usage{
merge_rows_omv(
  fleInp = c(),
  fleOut = "",
  typMrg = c("all", "common"),
  colInd = FALSE,
  rstRwN = TRUE,
  rmvDpl = FALSE,
  usePkg = c("haven", "foreign"),
  selSet = "",
  ...
)
}
\arguments{
\item{fleInp}{Vector with file names (including the path, if required) of the data files to be read (c("FILE1.omv", "FILE2.omv"); default: c())}

\item{fleOut}{Name of the data file to be written (including the path, if required; "FILE_OUT.omv"; default: ""); if empty, the data frame with the added columns is returned as variable (but not written)}

\item{typMrg}{Type of merging operation: "all" (default) or  "common"; see also Details}

\item{colInd}{Add a column with an indicator (the basename of the file minus the extension) marking from which input data set the respective rows are coming (default: FALSE)}

\item{rstRwN}{Reset row names (i.e., do not keep the row names of the original input data sets but number them consecutively - one to the row number of all input data sets added up; default: TRUE)}

\item{rmvDpl}{Remove duplicated rows (i.e., rows with the same content as a previous row in all columns; default: FALSE)}

\item{usePkg}{Name of the package: "haven" or "foreign" that shall be used to read SPSS, Stata and SAS files; "haven" is the default (it is more comprehensive), but with problems you may try "foreign"}

\item{selSet}{Name of the data set that is to be selected from the workspace (only applies when reading .Rdata-files)}

\item{...}{Additional arguments passed on to methods; see Details below}
}
\value{
a data frame (if fleOut is empty) with where the rows of all input data sets (i.e., the files given in the fleInp-argument) are concatenated
}
\description{
Merges two .omv-files for the statistical spreadsheet 'jamovi' (www.jamovi.org) by adding the content of the second file (fleIn2) as rows to the first file (fleIn1)
}
\details{
The different types of merging operations: "all" keeps all existing variables / columns that are contained in any of the input data sets and fills them up with NA where the variable / column doesn't
exist in a input data set. "common" only keeps the variables / columns that are common to all input data sets (i.e., that are contained in all data sets).
The ellipsis-parameter can be used to submit arguments / parameters to the functions that are used for merging or reading the data. The merging operation uses "rbind". When reading the data, the
functions are: "read_omv" (for jamovi-files), "read.table" (for CSV / TSV files; using similar defaults as "read.csv" for CSV and "read.delim" for TSV which both are based upon "read.table" but with
adjusted defaults for the respective file types), "readRDS" (for rds-files), "read_sav" (needs R-package "haven") or "read.spss" (needs R-package "foreign") for SPSS-files, read_dta ("haven") /
read.dta ("foreign") for Stata-files, read_sas ("haven") for SAS-data-files, and read_xpt ("haven") / read.xport ("foreign") for SAS-transport-files.
Please note that if the columns / variables have attributes (e.g., labels), the attributes of the first input data set (i.e., those contained in the first file given in the fleInp-argument) determine
the attributes of the output data set (i.e., the attributes of the second to the last input data set are disregarded).
}
\examples{
\dontrun{
library(jmvReadWrite);
}

}
