% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps2jmv.R
\name{sps2jmv}
\alias{sps2jmv}
\title{Converts SPPS-syntax (either from a SPSS-syntax file .sps or read from a SPSS-output-file .spv \link{using spv2sps}) into jamovi / jmv-syntax}
\usage{
sps2jmv(lstSPS = list(), fleSPS = "", fleSAV = "")
}
\arguments{
\item{list}{with SPSS-commands, the format is supposed to be the same as what is output from spv2sps (a list containing one command per entry, the command being a character vector, and the attribute 'datafile' with a character vector containing the position of the SPSS-data-file \link{incl. path}; if not found at the path, the file is searched in the current directory)}

\item{position}{of a SPSS-data-file \link{.sav; incl. path}}
}
\value{
list with jamovi / jmv-analysis-function-calls and the data set stored in the attribute 'dataset'
}
\description{
The function expects that either the parameter lstSPS \link{with the output from spv2sps; otherwise the list has to have identical format – see below} is used or that both fleSPS and fleSAV are given.
}
