% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spv2sps.R
\name{spv2sps}
\alias{spv2sps}
\title{Reads SPSS output files (.spv) and extracts the syntax from them}
\usage{
spv2sps(fleNme = "")
}
\arguments{
\item{fleNme}{name (incl. path) of the SPSS-output-file to be read ("FILENAME.spv"; default = "")}
}
\value{
list with SPSS commands (like what you store in SPSS syntax files - .sps)
}
\description{
Reads SPSS output files (.spv) and extracts the syntax from them
}
