% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmvReadWrite.R
\docType{package}
\name{jmvReadWrite}
\alias{jmvReadWrite}
\title{jmvReadWrite: A package to read and write .omv-files (jamovi statistical spreadsheet, www.jamovi.org).}
\description{
The jmvReadWrite package provides two functions:
jmvRead and jmvWrite.
}
\section{jmvReadWrite functions}{

The package jmvReadWrite reads and write the files that are produced of the statistical spreadsheet jamovi (www.jamovi.org).
As syntax from jamovi can also be directly used in R, having import / export routines for the data files permits an easy
transfer of analyses between jamovi and R.
}

