% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmvReadWrite.R
\docType{package}
\name{jmvReadWrite}
\alias{jmvReadWrite}
\title{jmvReadWrite: A package to read and write '.omv'-files (produced by the
statistical spreadsheet 'jamovi', www.jamovi.org).}
\description{
The jmvReadWrite package provides the following functions:
read_jmv
write_jmv
spv2sps
}
\section{jmvReadWrite functions}{

The free and open a statistical spreadsheet 'jamovi' (www.jamovi.org) aims
to make statistical analyses easy and intuitive. 'jamovi' produces syntax
that can directly be used in R (in connection with the R-package 'jmv').
Having import / export routines for the data files generated by 'jamovi'
('.omv') permits an easy transfer of analyses between 'jamovi' and R.
}

