% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_omv.R
\name{replace_omv}
\alias{replace_omv}
\title{Search values in .omv-files for the statistical spreadsheet 'jamovi' (\url{https://www.jamovi.org})}
\usage{
replace_omv(
  dtaInp = NULL,
  fleOut = "",
  rplLst = list(),
  whlTrm = TRUE,
  incNum = TRUE,
  incOrd = TRUE,
  incNom = TRUE,
  incID = TRUE,
  incCmp = TRUE,
  incRcd = TRUE,
  psvAnl = FALSE,
  ...
)
}
\arguments{
\item{dtaInp}{Either a data frame or the name of a jamovi data file to be read (including the path, if required; "FILENAME.omv"; default: NULL)}

\item{fleOut}{Name of the data file to be written (including the path, if required; "FILE_OUT.omv"; default: ""); if empty, the resulting data frame is
returned instead}

\item{rplLst}{A list where each entry is a vector (with length 2) containing the original value and the to-replace-value (default: list())}

\item{whlTrm}{Whether the search term (first entry in the vectors) must be found exactly (TRUE) or whether a partial match is sufficient (FALSE; default:
TRUE)}

\item{incNum}{Whether to include continuous variables in the search (default: TRUE)}

\item{incOrd}{Whether to include ordinal variables in the search (default: TRUE)}

\item{incNom}{Whether to include nominal variables in the search (default: TRUE)}

\item{incID}{Whether to include ID variables in the search (default: TRUE)}

\item{incCmp}{Whether to include Computed variables in the search (default: TRUE)}

\item{incRcd}{Whether to include Recoded variables in the search (default: TRUE)}

\item{psvAnl}{Whether analyses that are contained in the input file shall be transferred to the output file (default: FALSE)}

\item{...}{Additional arguments passed on to methods; see Details below}
}
\value{
a data frame (only returned if \code{fleOut} is empty) with the replaced values
}
\description{
Search values in .omv-files for the statistical spreadsheet 'jamovi' (\url{https://www.jamovi.org})
}
\details{
\itemize{
\item \code{rplLst} is a list. Each list entry contains a vector (with length 2), where the first entry is the original value, and the second entry is the value the
original value is to be replaced with.
\item ``
\item The ellipsis-parameter (\code{...}) can be used to submit arguments / parameters to the function that is used for writing the data. Clicking on the respective
function under “See also”, you can get a more detailed overview over which parameters each of those functions take. The functions are:
\code{read_omv} and \code{write_omv} (for jamovi-files), \code{read.table} (for CSV / TSV files; using similar defaults as \code{read.csv} for CSV and \code{read.delim} for TSV
which both are based upon \code{read.table}), \code{load} (for .RData-files), \code{readRDS} (for .rds-files), \code{read_sav} (needs the R-package \code{haven}) or \code{read.spss}
(needs the R-package \code{foreign}) for SPSS-files, \code{read_dta} (\code{haven}) / \code{read.dta} (\code{foreign}) for Stata-files, \code{read_sas} (\code{haven}) for SAS-data-files,
and \code{read_xpt} (\code{haven}) / \code{read.xport} (\code{foreign}) for SAS-transport-files. If you would like to use \code{haven}, you may need to install it using
\code{install.packages("haven", dep = TRUE)}.
}
}
\examples{
\dontrun{
library(jmvReadWrite)
rplLst = list(c(), c())
}

}
\seealso{
\code{replace_omv} uses \code{\link[=write_omv]{write_omv()}} for writing jamovi-files.
}
