% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_omv.R
\name{search_omv}
\alias{search_omv}
\title{Search values in .omv-files for the statistical spreadsheet 'jamovi' (\url{https://www.jamovi.org})}
\usage{
search_omv(
  dtaInp = NULL,
  srcTrm = c(),
  whlTrm = FALSE,
  incNum = TRUE,
  incOrd = TRUE,
  incNom = TRUE,
  incID = TRUE,
  incCmp = TRUE,
  incRcd = TRUE,
  ...
)
}
\arguments{
\item{dtaInp}{Either a data frame or the name of a jamovi data file to be read (including the path, if required; "FILENAME.omv"; default: NULL)}

\item{srcTrm}{(Character or numeric) Vector (with length = 1) with a search term to be found in the data frame (default: c())}

\item{whlTrm}{Whether the exact search term shall be found (TRUE) or whether a partial match is sufficient (FALSE; default: FALSE)}

\item{incNum}{Whether to include continuous variables in the search (default: TRUE)}

\item{incOrd}{Whether to include ordinal variables in the search (default: TRUE)}

\item{incNom}{Whether to include nominal variables in the search (default: TRUE)}

\item{incID}{Whether to include ID variables in the search (default: TRUE)}

\item{incCmp}{Whether to include Computed variables in the search (default: TRUE)}

\item{incRcd}{Whether to include Recoded variables in the search (default: TRUE)}

\item{...}{Additional arguments passed on to methods; see Details below}
}
\value{
a named list with the places where the search term was found (names in the list are the variables / columns, the entries the respective row
numbers within that variable / column)
}
\description{
Search values in .omv-files for the statistical spreadsheet 'jamovi' (\url{https://www.jamovi.org})
}
\examples{
\dontrun{
library(jmvReadWrite)
nmeInp <- system.file("extdata", "AlbumSales.omv", package = "jmvReadWrite")
}

}
