% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps2jmv.R
\name{sps2jmv}
\alias{sps2jmv}
\title{Converts SPPS-syntax (either from a SPSS-syntax file .sps or read from a SPSS-output-file .spv \link{using spv2sps}) into jamovi / jmv-syntax}
\usage{
sps2jmv(vecSPS = c(), fleSPS = "", fleSAV = "", fltCnd = "")
}
\arguments{
\item{character}{vector with SPSS-commands, the format is supposed to be the same as what is output from spv2sps: a character vector containing one command per element, and the attribute "datafile" that contains a string pointing to the location of the SPSS-data-file \link{incl. path} (if not found at the path or if no path is given, the file is searched in the current directory)}

\item{string}{a filter condition to be applied to the dataset (default = "" → no filter; e.g., "data$\link{VARNAME} == 1")}
}
\value{
list with jamovi / jmv-analysis-function-calls and the data set stored in the attribute "dataset"
}
\description{
The function expects that either the parameter vecSPS is used \link{with the output from spv2sps} or that both fleSPS and fleSAV are given.
}
