% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToothGrowth}
\alias{ToothGrowth}
\title{The Effect of Vitamin C on Tooth Growth in Guinea Pigs}
\format{
A data.frame with 60 rows and 6 variables
\describe{
\item{logLen}{numeric}{Natural logarithm of the tooth length (len)}
\item{supp - Transform 1}{factor}{Transformation of the supplement type (factor to numerical: VC = 1; OJ = 2)}
\item{len}{numeric}{Tooth length}
\item{supp}{factor}{Supplement type (VC: Vitamin C or OJ: Orange juice)}
\item{dose}{numeric}{Dose in grams / day}
}
}
\usage{
ToothGrowth
}
\description{
The Effect of Vitamin C on Tooth Growth in Guinea Pigs
}
\keyword{datasets}
