% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_omv.R
\name{write_omv}
\alias{write_omv}
\title{Write files to be used with the statistical spreadsheet 'jamovi'
(www.jamovi.org)}
\usage{
write_omv(dtaFrm = NULL, fleNme = "")
}
\arguments{
\item{dtaFrm}{Data frame to be exported (default: NULL)}

\item{fleNme}{Name / position of the output file to be generated ("FILENAME.omv"; default: "")}
}
\value{
a list containing the meta data (mtaDta, written to metadata.json in the OMV-file), the extended data (xtdDta, written to xdata.json in the OMV-file) and the original data frame (dtaFrm)
}
\description{
Write files to be used with the statistical spreadsheet 'jamovi'
(www.jamovi.org)
}
\examples{
library(jmvReadWrite)

# use the data set "ToothGrowth" and write it as jamovi-file using write_omv()
data("ToothGrowth");
wrtDta = write_omv(ToothGrowth, "Trial.omv");

names(wrtDta);
# → "mtaDta" "xtdDta" "dtaFrm"
# returns a list with the metadata (mtaDta, e.g., column and data type),
# the extended data (xtdDta, e.g., variable lables), and the data frame (dtaFrm)
# the purpose of these variables is merely for checking (understanding the file format)
# and debugging

# check whether the file was written to the disk, get the file information (size, etc.)
# and delete the file afterwards
list.files(".", "Trial.omv");
file.info("Trial.omv");
unlink("Trial.omv");

}
