% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_omv.R
\name{write_omv}
\alias{write_omv}
\title{Write files to be used with the statistical spreadsheet 'jamovi'
(www.jamovi.org)}
\usage{
write_omv(dtaFrm = NULL, fleNme = "")
}
\arguments{
\item{dtaFrm}{Data frame to be exported (default = NULL)}

\item{fleNme}{Name / position of the output file to be generated ("FILENAME.omv"; default = "")}
}
\description{
Write files to be used with the statistical spreadsheet 'jamovi'
(www.jamovi.org)
}
